Fixed misspelling + content-type
when i tried my first test it didnt work, i saw that i had a spelling mistake so i quickly fixed it. Even though i fixed it, it still seem to work. Therefor i checked the problem; saw that it was CORS, and began to try different methods to fix it. I added new headers to my headersCORS but that didnt do anything, then i tired to set the headersCORS in the GET-if statement to a different content-type, this worked. In the code i never added the actual CORS solving header, so it wont include that header. So instead i decided to set the content-type and then add the whole object